% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_delta.R
\name{boots_delta}
\alias{boots_delta}
\title{Title}
\usage{
boots_delta(
  A_est,
  Pi_est,
  x_from,
  prob1,
  h,
  Sel_from,
  al,
  seuil,
  min_size,
  n,
  max_pi0,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  type_init,
  approx,
  maxit = 100,
  f0_known = TRUE,
  delta,
  type_check = "closer_to0"
)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
\examples{
m <- 2000
theta <- sim_markov(m, Pi = c(0.8,0.2), A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T))
x <- rep(0, m)
x[theta == 0] <- rnorm(sum(theta ==0))
x[theta == 1] <- rnorm(sum(theta ==1), 2, 1)
Est <- Estimation(x, m0_init = 0, sd0_init = 1, norm_init = TRUE, plot= TRUE)
Sel <- Selection_delta(x, Est$Em$fw_bc_EM, 0.05,
 Est$Em$A,  Est$Em$f0x,  Est$Em$f1x,
  Est$Em$Pi, 9)
boots_delta(Est$Em$A, Est$Em$Pi,
x_from = x,
prob1 = Est$Em$fw_bc_EM$gamma[,2], h,
Sel, al =0.1,
seuil=0.05,
min_size =9,b_act*s_dbnr,
n = n,
max_pi0 = 0.99999,
m0_init =  0, sd0_init = 1,
 norm_init = TRUE,
type_init = "given",
approx = TRUE,
delta = 0.9)
}
