% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borne.R
\name{borne}
\alias{borne}
\title{Title}
\usage{
borne(type_borne, sel, a, alpha, m, pval, C, ZL, leaf_list)
}
\arguments{
\item{type_borne}{}

\item{sel}{}

\item{a}{}

\item{alpha}{}

\item{m}{}

\item{pval}{}

\item{C}{}

\item{ZL}{}

\item{leaf_list}{}
}
\description{
Title
}
