% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Estimation}
\alias{Estimation}
\title{Title}
\usage{
Estimation(
  x,
  h = 0.3,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  f0_known = TRUE,
  f0x_est = NULL,
  pval = NULL,
  plot = FALSE,
  size_plot = min(10000, length(x)),
  approx = TRUE,
  maxit = 1000
)
}
\arguments{
\item{x}{value of the statistics}

\item{h}{window for the kernel estimation}

\item{m0_init}{if norm_init TRUE, initial expectency of the law under H0}

\item{sd0_init}{if norm_init TRUE, initial sd of the law under H0}

\item{df_init}{degree of freedom if f0 is initialized with a student}

\item{norm_init}{either f0 is initialized with a normal distribution}

\item{max_pi0}{pi0 cann't get to close to one it is the maximum value}

\item{f0_known}{wether or not  f0 is known. (If not f0 is estimated)}

\item{f0x_est}{(if f0 is known it's the value of f0 for each element of x)}

\item{pval}{one can also provide directly some p-value}

\item{plot}{represents the density ?}

\item{size_plot}{}

\item{approx}{}

\item{maxit}{maximum number of iteration}
}
\value{
a list with the estimated parameters (the value of f0 and f1 for all element of x and the transition matrix A)
}
\description{
Title
}
