% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantile.R
\name{get_IC}
\alias{get_IC}
\title{Title}
\usage{
get_IC(sel, li0, Pis, f0x, f1x, alpha)
}
\arguments{
\item{sel}{must be ordered!}

\item{li0}{}

\item{Pis}{}

\item{f0x}{}

\item{f1x}{}
}
\description{
Title
}
\examples{
 m <-  100
 Pi <- c(0.8,0.2)
 A <- matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T)
   rdata <- sim_hmm_2states(m, Pi, A, f0 = c(0,1), f1= c(2,1))
  x <- rdata$x
  f0x <- dnorm(x)
  f1x <- dnorm(x, 1,2)
   mod <- for_back(m, A, f0x, f1x, Pi)
     Pis_est <- lapply(2:m, function(i){
  get_A( m,alpha = mod$alpha, beta = mod$beta, A, f0x, 
         f1x, i = i)
  })
  alpha <- 0.1
  sel <- sample(1:m, m/2)
  sel <- sel[order(sel)]
 get_IC(sel = sel, li0 = mod$gamma[, 1], 
  Pis = Pis_est, f0x = f0x, f1x = f1x, alpha)
 quant <- get_quantiles(sel = sel, li0 = mod$gamma[, 1], 
  Pis = Pis_est, f0x = f0x, f1x = f1x)
  borne(type_borne = "HMM", sel = sel, a = quant, alpha )
  borne(type_borne = "HMM_small", sel = sel, a = quant, alpha)
}
