% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantile.R
\name{get_probs}
\alias{get_probs}
\title{Title}
\usage{
get_probs(sel, li0, Pis, f0x, f1x, probs)
}
\arguments{
\item{sel}{}

\item{li0}{}

\item{Pis}{}

\item{f0x}{}

\item{f1x}{}

\item{probs}{}
}
\description{
Title
}
