% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantile.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Title}
\usage{
get_quantiles(sel, li0, Pis, f0x, f1x)
}
\arguments{
\item{sel}{}

\item{li0}{}

\item{Pis}{}

\item{f0x}{}

\item{f1x}{}
}
\description{
Title
}
