% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_delta.R
\name{Selection_delta}
\alias{Selection_delta}
\title{Title}
\usage{
Selection_delta(
  x,
  fw_bc,
  seuil,
  A_est,
  f0x_est,
  f1x_est,
  Pi_est,
  min_size,
  min_jump = NULL,
  pval = NULL,
  all = FALSE
)
}
\arguments{
\item{x}{}

\item{fw_bc}{}

\item{seuil}{}

\item{A_est}{}

\item{f0x_est}{}

\item{f1x_est}{}

\item{Pi_est}{}

\item{min_size}{}

\item{min_jump}{}

\item{pval}{}

\item{all}{}
}
\description{
Title
}
