% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_tibble.R
\name{Selection_tibble}
\alias{Selection_tibble}
\title{Title}
\usage{
Selection_tibble(
  x,
  fw_bc,
  seuil,
  A_est,
  f0x_est,
  f1x_est,
  Pi_est,
  min_size,
  min_jump = NULL,
  pval = NULL,
  all = FALSE
)
}
\arguments{
\item{x}{}

\item{fw_bc}{}

\item{seuil}{}

\item{A_est}{}

\item{f0x_est}{}

\item{f1x_est}{}

\item{Pi_est}{}

\item{pval}{}
}
\description{
Title
}
