% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utile.R
\name{sim_hmm_from_weightkde}
\alias{sim_hmm_from_weightkde}
\title{Title}
\usage{
sim_hmm_from_weightkde(A, Pi, x_from, prob1, h, n)
}
\arguments{
\item{A}{}

\item{Pi}{}

\item{x_from}{}

\item{prob1}{}

\item{h}{}

\item{n}{}
}
\description{
Title
}
