% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_delta.R
\name{simu_delta}
\alias{simu_delta}
\title{Title}
\usage{
simu_delta(
  m,
  A,
  Pi,
  n,
  rho,
  SNR,
  prob,
  type_sim = "HMM",
  al,
  s_dbnr,
  b_act,
  d,
  seuil,
  h = 0.3,
  n_boot,
  min_size,
  norm,
  m0,
  sd0,
  df,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  type_init,
  num_seed,
  f0_known,
  approx,
  all = FALSE,
  size_b0 = 300,
  pct_b1 = 1/3,
  include_H0 = FALSE,
  delta,
  n_seg,
  drop1,
  drop2,
  tumorFraction,
  sim_markov_nonstat = NULL,
  sel_function = Selection_delta
)
}
\arguments{
\item{delta}{}
}
\description{
Title
}
\examples{
simu_delta( m = c(100),
A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T),
Pi = c(0.95, 0.05),
n = c(500),
rho = c(0),
SNR = 2,
prob = c(0.5),
type_sim = c("HMM"),
n_boot = 20,
al = 0.2, s_dbnr = 10, b_act = 2, d = 1, seuil= 0.05,
min_size = 2, norm = TRUE, sd0 = 0.5, m0= 0,sd0_init = 0.5, m0_init= 0,
norm_init = TRUE, df= 2, num_seed= 1234, type_init="given", f0_known=TRUE,
approx = TRUE, delta = 0.9)
}
