% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{simu_delta2}
\alias{simu_delta2}
\title{Title}
\usage{
simu_delta2(
  m,
  A,
  Pi,
  n,
  rho,
  SNR,
  prob,
  type_sim = "HMM",
  al,
  s_dbnr,
  b_act,
  d,
  seuil,
  h = 0.3,
  n_boot,
  min_size,
  norm,
  m0,
  sd0,
  df,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  type_init,
  num_seed,
  f0_known,
  approx,
  all = FALSE,
  size_b0 = 300,
  pct_b1 = 1/3,
  include_H0 = FALSE,
  delta,
  n_seg,
  drop1,
  drop2,
  tumorFraction
)
}
\arguments{
\item{m}{}

\item{A}{}

\item{Pi}{}

\item{n}{}

\item{rho}{}

\item{SNR}{}

\item{prob}{}

\item{type_sim}{}

\item{al}{}

\item{s_dbnr}{}

\item{b_act}{}

\item{d}{}

\item{seuil}{}

\item{h}{}

\item{n_boot}{}

\item{min_size}{}

\item{norm}{}

\item{m0}{}

\item{sd0}{}

\item{df}{}

\item{m0_init}{}

\item{sd0_init}{}

\item{df_init}{}

\item{norm_init}{}

\item{max_pi0}{}

\item{type_init}{}

\item{num_seed}{}

\item{f0_known}{}

\item{approx}{}

\item{all}{}

\item{size_b0}{}

\item{pct_b1}{}

\item{include_H0}{}

\item{delta}{}

\item{n_seg}{}

\item{drop1}{}

\item{drop2}{}

\item{tumorFraction}{}
}
\value{

}
\description{
Title
}
