% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{Estimation}
\alias{Estimation}
\title{Title}
\usage{
Estimation(
  x,
  h = 0.3,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  f0_known = TRUE,
  f0x_est = NULL,
  pval = NULL,
  plot = FALSE,
  size_plot = min(10000, length(x)),
  approx = TRUE,
  maxit = 1000
)
}
\arguments{
\item{x}{}

\item{h}{}

\item{m0_init}{}

\item{sd0_init}{}

\item{df_init}{}

\item{norm_init}{}

\item{max_pi0}{}

\item{f0_known}{}

\item{f0x_est}{}

\item{pval}{}

\item{plot}{}

\item{size_plot}{}

\item{approx}{}

\item{maxit}{}
}
\description{
Title
}
