% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_delta.R, R/simulation.R
\name{simu_delta}
\alias{simu_delta}
\title{Title}
\usage{
simu_delta(
  m,
  A,
  Pi = NULL,
  rho,
  SNR,
  prob,
  type_sim = "HMM",
  al,
  s_dbnr,
  b_act,
  d,
  seuil,
  h = 0.3,
  n_boot,
  min_size,
  norm,
  m0,
  sd0,
  df,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  type_init,
  num_seed,
  f0_known,
  approx,
  all = FALSE,
  size_b0 = 300,
  pct_b1 = 1/3,
  include_H0 = FALSE,
  delta,
  n_seg,
  drop1,
  drop2,
  tumorFraction,
  sel_function = Selection_delta
)

simu_delta(
  m,
  A,
  Pi = NULL,
  rho,
  SNR,
  prob,
  type_sim = "HMM",
  al,
  s_dbnr,
  b_act,
  d,
  seuil,
  h = 0.3,
  n_boot,
  min_size,
  norm,
  m0,
  sd0,
  df,
  m0_init,
  sd0_init,
  df_init,
  norm_init,
  max_pi0 = 0.99999,
  type_init,
  num_seed,
  f0_known,
  approx,
  all = FALSE,
  size_b0 = 300,
  pct_b1 = 1/3,
  include_H0 = FALSE,
  delta,
  n_seg,
  drop1,
  drop2,
  tumorFraction,
  sel_function = Selection_delta
)
}
\arguments{
\item{m}{numeric, the number of hypothesis}

\item{A}{matrix, the transition matrix of the markov chain}

\item{Pi}{a numeric vector, the initial distribution (if different from the stationary distribution)}

\item{SNR}{numeric, the Signal to noise ratio}

\item{type_sim}{character, the type of simulation made}

\item{al}{numeric (between 0 and 1),  the risk}

\item{seuil}{numeric, the threshold for the pvalues}

\item{h}{numeric, the window in kernel density estimation}

\item{n_boot}{numeric,  the number of bootstrap samples}

\item{min_size}{numeric,  the minimum size of a block of consecutive selected hypothesis}

\item{norm}{logical, whether densities are gaussian or not (otherwise it is assume to be a student)}

\item{m0}{numeric, the expectency under the null}

\item{sd0}{numeric,  the standard deviation under the null}

\item{df}{numeric, the degres of freedom}

\item{m0_init}{numeric, if f0 is unknown the initialisation of its expectency}

\item{sd0_init}{numeric, if f0 is unknown the initialisation of its expectency}

\item{df_init}{numeric, if f0 is unknown the initialisation of its expectency}

\item{norm_init}{logical, if f0 is unknown wether the initialisation of its law is gaussian}

\item{max_pi0}{numeric, maximum value of pi0}

\item{type_init}{}

\item{num_seed}{to set a seed.}

\item{f0_known}{logical, wether f0 is assumed to be known or not}

\item{approx}{logical wether to estimate the density at every point or to add a small linear approximation between a large range of point}

\item{all}{logical, compare all the}

\item{size_b0}{numerical, size of continuous block for type_sim = "block"}

\item{pct_b1}{numerical, ratio of the size of the block une H1 / size_b0}

\item{include_H0}{logical, wether to add the information on H0 or not}

\item{delta}{numeric, the value of delta (the part of the risk for the bootstrap and the other one)}
}
\description{
Title

Title
}
\examples{
simu_delta( m = c(100),
A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T),
Pi = c(0.95, 0.05),
n = c(500),
rho = c(0),
SNR = 2,
prob = c(0.5),
type_sim = c("HMM"),
n_boot = 20,
al = 0.2, s_dbnr = 10, b_act = 2, d = 1, seuil= 0.05,
min_size = 2, norm = TRUE, sd0 = 0.5, m0= 0,sd0_init = 0.5, m0_init= 0,
norm_init = TRUE, df= 2, num_seed= 1234, type_init="given", f0_known=TRUE,
approx = TRUE, delta = 0.9)
simu_delta( m = c(100),
A = matrix(c(0.95, 0.05, 0.2, 0.80), 2, 2, byrow = T),
Pi = c(0.95, 0.05),
rho = c(0),
SNR = 2,
prob = c(0.5),
type_sim = c("HMM"),
n_boot = 20,
al = 0.2, s_dbnr = 10, b_act = 2, d = 1, seuil= 0.05,
min_size = 2, norm = TRUE, sd0 = 0.5, m0= 0,sd0_init = 0.5, m0_init= 0,
norm_init = TRUE, df= 2, num_seed= 1234, type_init="given", f0_known=TRUE,
approx = TRUE, delta = 0.9)
}
